pyproject.toml
[project]
name = "friday"
version = "0.1.0"
description = "LLM-powered shell agent — a ZSH extension for coding, debugging, and shell assistance"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.92,<1",
    "httpx>=0.28,<1",
    "openai>=2.31,<3",
    "pydantic>=2.12,<3",
    "pydantic-ai>=1.78,<1.79",
    "pydantic-settings>=2.13,<3",
    "typer>=0.15",
    "rich>=13.0",
    "prompt-toolkit>=3.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]

[project.scripts]
friday = "friday.cli.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/friday"]
artifacts = ["src/friday/agent/prompts/*.md"]

[dependency-groups]
dev = ["ruff>=0.11", "ty>=0.0.26", "pre-commit>=4.0", "taskipy>=1.14"]

test = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "pytest-asyncio>=0.25",
    "pytest-xdist>=3.0",
    "pytest-mock>=3.0",
]

[tool.ruff]
src = ["src", "tests"]
extend-exclude = [".claude/"]
line-length = 100
target-version = "py313"
fix = true

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "S", "RUF"]
ignore = ["S101"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S"]
"src/friday/agent/context.py" = ["S603", "S607"]  # git subprocess calls
"src/friday/tools/filesystem.py" = ["S603"]        # ripgrep/grep calls
"src/friday/tools/shell.py" = ["S602"]             # shell=True is the tool's purpose

[tool.ruff.format]
quote-style = "single"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = ["-v", "--strict-markers", "--tb=short"]
markers = [
    "slow: long-running tests",
    "integration: integration-level tests",
    "e2e: end-to-end flow tests",
]

[tool.coverage.run]
source = ["src"]
branch = true
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
show_missing = true
skip_covered = false

[tool.taskipy.tasks]
format = "ruff format ."
lint = "ruff check . --fix"
typecheck = "ty check --exclude 'tests/'"
test = "pytest -v"
check = "task format && task lint && task typecheck && task test"
